feat: refactor total balance calculation to use useTotalBalanceByAccount hook
